fix(climc): optimized storagecache cmd
